refactor: 템플릿 메서드 적용, user 의존도 제거
