--- jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -1,9 +1,0 @@
-package nextstep.jdbc;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class JdbcTemplate {
-
-    private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
-}
